backend/app/ai/face_recognition_engine.py
import base64
import numpy as np
import io
import os
import time
from PIL import Image
from sqlalchemy.orm import Session
from datetime import datetime, timezone

try:
    import dlib
    DLIB_AVAILABLE = True
except ImportError:
    DLIB_AVAILABLE = False
    print("WARNING: dlib not installed. Face recognition disabled.")

from app.models.user import User
from app.models.attendance import FaceRecognitionLog

# ─── Model Paths ────────────────────────────────────────────────────
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
MODELS_DIR = os.path.join(BASE_DIR, "models")

LANDMARK_MODEL = os.path.join(MODELS_DIR, "shape_predictor_68_face_landmarks.dat")
RECOGNITION_MODEL = os.path.join(MODELS_DIR, "dlib_face_recognition_resnet_model_v1.dat")


class FaceRecognitionEngine:
    """
    Dlib-based Face Recognition Engine.
    128-d float64 embeddings binary blob format mein DB mein store karta hai.
    """

    CONFIDENCE_THRESHOLD = 0.50   # distance <= 0.5 = match (dlib standard)

    # ─── Lazy-load dlib models ──────────────────────────────────────
    _detector = None
    _predictor = None
    _face_rec = None

    @classmethod
    def _load_models(cls):
        if not DLIB_AVAILABLE:
            return False
        if cls._detector is not None:
            return True

        if not os.path.exists(LANDMARK_MODEL):
            print(f"ERROR: Landmark model not found at {LANDMARK_MODEL}")
            return False
        if not os.path.exists(RECOGNITION_MODEL):
            print(f"ERROR: Recognition model not found at {RECOGNITION_MODEL}")
            return False

        cls._detector  = dlib.get_frontal_face_detector()
        cls._predictor = dlib.shape_predictor(LANDMARK_MODEL)
        cls._face_rec  = dlib.face_recognition_model_v1(RECOGNITION_MODEL)
        print("✅ Dlib models loaded successfully")
        return True

    # ─── Helpers ────────────────────────────────────────────────────

    @staticmethod
    def _decode_image(image_base64: str) -> np.ndarray:
        """Base64 string → RGB numpy array"""
        # Data-URL prefix hata do agar hai
        if "," in image_base64:
            image_base64 = image_base64.split(",", 1)[1]
        image_data = base64.b64decode(image_base64)
        image = Image.open(io.BytesIO(image_data)).convert("RGB")
        return np.array(image)

    @staticmethod
    def _embedding_to_bytes(embedding: np.ndarray) -> bytes:
        return embedding.astype(np.float64).tobytes()

    @staticmethod
    def _bytes_to_embedding(embedding_bytes: bytes) -> np.ndarray:
        return np.frombuffer(embedding_bytes, dtype=np.float64)

    @classmethod
    def _extract_embedding(cls, image_array: np.ndarray):
        """
        Image se 128-d embedding nikalo.
        Returns: (embedding np.ndarray, error_str or None)
        """
        if not cls._load_models():
            return None, "Dlib models not loaded"

        # dlib ko uint8 RGB chahiye
        if image_array.dtype != np.uint8:
            image_array = image_array.astype(np.uint8)

        detections = cls._detector(image_array, 1)

        if len(detections) == 0:
            return None, "No face detected in image"
        if len(detections) > 1:
            return None, "Multiple faces detected — use single face image"

        face = detections[0]
        shape = cls._predictor(image_array, face)
        descriptor = cls._face_rec.compute_face_descriptor(image_array, shape)
        embedding = np.array(descriptor)  # 128-d float64
        return embedding, None

    @staticmethod
    def _euclidean_distance(a: np.ndarray, b: np.ndarray) -> float:
        return float(np.linalg.norm(a - b))

    # ─── Public API ─────────────────────────────────────────────────

    @classmethod
    def enroll_face(cls, db: Session, user_id: int, image_base64: str) -> dict:
        """
        Kisi bhi user (student/teacher/admin) ka face enroll karo.
        """
        if not cls._load_models():
            return {"success": False, "error": "Dlib models not available"}

        try:
            start = time.time()

            image_array = cls._decode_image(image_base64)
            embedding, error = cls._extract_embedding(image_array)

            if error:
                return {"success": False, "error": error}

            user = db.query(User).filter(User.id == user_id).first()
            if not user:
                return {"success": False, "error": "User not found"}

            user.face_embedding = cls._embedding_to_bytes(embedding)
            user.face_enrolled_at = datetime.now(timezone.utc)
            db.commit()

            processing_time = int((time.time() - start) * 1000)

            return {
                "success": True,
                "user_id": user_id,
                "processing_time_ms": processing_time,
                "enrolled_at": str(user.face_enrolled_at),
            }

        except Exception as e:
            return {"success": False, "error": str(e)}

    @classmethod
    def login_by_face(cls, db: Session, image_base64: str) -> dict:
        """
        Face se login karo.
        Saare enrolled users se compare karo, best match return karo.
        Returns: {success, user_id, role, confidence, full_name, ...}
        """
        if not cls._load_models():
            return {"success": False, "error": "Dlib models not available"}

        try:
            start = time.time()

            image_array = cls._decode_image(image_base64)
            query_embedding, error = cls._extract_embedding(image_array)

            if error:
                return {"success": False, "error": error}

            # Saare enrolled users load karo
            users = db.query(User).filter(
                User.face_embedding.isnot(None),
                User.is_active == True
            ).all()

            if not users:
                return {"success": False, "error": "No enrolled faces found in system"}

            best_user = None
            best_distance = float("inf")

            for user in users:
                try:
                    stored_emb = cls._bytes_to_embedding(user.face_embedding)
                    if len(stored_emb) != 128:
                        continue
                    dist = cls._euclidean_distance(query_embedding, stored_emb)
                    if dist < best_distance:
                        best_distance = dist
                        best_user = user
                except Exception:
                    continue

            processing_time = int((time.time() - start) * 1000)

            if best_user is None or best_distance > cls.CONFIDENCE_THRESHOLD:
                return {
                    "success": False,
                    "error": "Face not recognized",
                    "confidence": round(1 - best_distance, 3) if best_user else 0,
                    "processing_time_ms": processing_time,
                }

            # Confidence = 1 - distance (0 to 1)
            confidence = round(1.0 - best_distance, 3)

            return {
                "success": True,
                "user_id": best_user.id,
                "role": best_user.role,
                "email": best_user.email,
                "confidence": confidence,
                "distance": round(best_distance, 4),
                "processing_time_ms": processing_time,
            }

        except Exception as e:
            return {"success": False, "error": str(e)}

    @classmethod
    def recognize_face(
        cls,
        db: Session,
        image_base64: str,
        gate_id: int = None,
        camera_id: int = None
    ) -> dict:
        """
        Attendance ke liye face recognize karo (backward compat).
        """
        result = cls.login_by_face(db, image_base64)

        if not result["success"]:
            if gate_id and camera_id:
                cls._log_attempt(db, None, gate_id, camera_id, 0, False, result.get("error"))
            return {
                "matched": False,
                "confidence": result.get("confidence", 0),
                "error": result.get("error"),
                "processing_time_ms": result.get("processing_time_ms", 0),
            }

        if gate_id and camera_id:
            cls._log_attempt(
                db,
                result["user_id"],
                gate_id,
                camera_id,
                result["confidence"],
                True,
                None,
                result["processing_time_ms"],
            )

        return {
            "matched": True,
            "student_id": result["user_id"],
            "user_id": result["user_id"],
            "confidence": result["confidence"],
            "processing_time_ms": result["processing_time_ms"],
            "spoof_check_passed": True,
            "liveness_score": 0.9,
        }

    @staticmethod
    def _log_attempt(
        db: Session,
        user_id,
        gate_id: int,
        camera_id: int,
        confidence: float,
        success: bool,
        error_msg: str = None,
        processing_time_ms: int = 0,
    ):
        try:
            log = FaceRecognitionLog(
                student_id=user_id,
                gate_id=gate_id,
                camera_id=camera_id,
                confidence=confidence,
                match_success=success,
                processing_time_ms=processing_time_ms,
                spoof_check_passed=True,
                liveness_score=0.9,
                error_message=error_msg,
            )
            db.add(log)
            db.commit()
        except Exception:
            pass