backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
starlette==0.38.6

pydantic==2.9.2
pydantic-settings==2.6.1
typing_extensions==4.15.0

bcrypt==4.1.3
passlib==1.7.4
python-jose==3.3.0

SQLAlchemy==2.0.36
greenlet==3.3.2
PyMySQL==1.1.1

python-dotenv==1.0.1
python-multipart==0.0.12
email-validator==2.3.0

httpx==0.27.2

numpy==1.26.4
scikit-learn==1.5.2
scipy==1.11.4
joblib==1.5.3

matplotlib==3.8.4
pillow==10.4.0

dlib==20.0.0
dlib-bin==20.0.0

cryptography==46.0.5